refactor(product-tailoring): restructure models and exports
