+++ models/product-tailoring/src/product-tailoring-data/product-tailoring-data-draft/fields-config.ts
